fix(database): refactor file structure for visibility, added lib components where possible
