--- apps/Conduit-UI/src/components/database/CmsModels.ts
@@ -1,14 +1,0 @@
-export interface SchemaUI {
-  createdAt: string;
-  name: string;
-  updatedAt: string;
-  _id: string;
-  modelOptions: {
-    conduit: {
-      cms: {
-        authentication: boolean;
-        crudOperations: boolean;
-      };
-    };
-  };
-}
