+++ apps/Conduit-UI/src/components/database/schemas/EditSchema/BuildTypesContent.tsx
@@ -6,12 +6,12 @@ import Typography from '@mui/material/Typography';
 import DeleteIcon from '@mui/icons-material/Delete';
 import SettingsIcon from '@mui/icons-material/Settings';
 import DragHandleIcon from '@mui/icons-material/DragHandle';
-import SimpleType from './types/SimpleType/SimpleType';
-import BooleanType from './types/BooleanType/BooleanType';
-import GroupType from './types/GroupType/GroupType';
-import EnumType from './types/EnumType/EnumType';
-import ObjectIdType from './types/ObjectIdType/ObjectIdType';
-import RelationType from './types/RelationType/RelationType';
+import SimpleType from '../../types/SimpleType/SimpleType';
+import BooleanType from '../../types/BooleanType/BooleanType';
+import GroupType from '../../types/GroupType/GroupType';
+import EnumType from '../../types/EnumType/EnumType';
+import ObjectIdType from '../../types/ObjectIdType/ObjectIdType';
+import RelationType from '../../types/RelationType/RelationType';
 import Button from '@mui/material/Button';
 import { BoxProps } from '@mui/material/Box/Box';
 import { styled } from '@mui/material';
