+++ apps/Conduit-UI/src/components/database/schemas/SchemaData/SchemaDataHeader.tsx
@@ -6,7 +6,7 @@ import { Paginator } from 'ui-components';
 import { BoxProps } from '@mui/material/Box/Box';
 import InputAdornment from '@mui/material/InputAdornment';
 import { Search, Refresh, AccountTree } from '@mui/icons-material';
-import useParseQuery from './useParseQuery';
+import useParseQuery from '../useParseQuery';
 import { styled, Typography } from '@mui/material';
 
 const ObjText = '{ }';
