+++ apps/Conduit-UI/src/components/database/types/EnumType/EnumType.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import Box from '@mui/material/Box';
-import FieldIndicators from '../../FieldIndicators';
+import { FieldIndicators } from 'ui-components';
 import SelectIcon from '@mui/icons-material/FormatListBulleted';
 import Grid from '@mui/material/Grid';
 import { CustomIcon } from '../SimpleType/SimpleType';
